bum version number add new macro PA_API_VERSION for preprocessor level conditional compiling add new native APIs: 	- counter 	- cork & flush for record streams 	- add flags parameters to pa_stream_connect_xx() 	- new prebuf command 	- time api, and total latency calculator 	- return sample spec ability to cork source output streams dump server status on SIGHUP to syslog show sink input/source outputs status in cli-text.c don't flush esound output buffer when client disconnects move version api to polyplib-version.h
git-svn-id: file:///home/lennart/svn/public/pulseaudio/trunk@240 fefdeb5f-60dc-0310-8127-8f9354f1896f

+++ polyp/sample.c
@@ -59,7 +59,7 @@ size_t pa_bytes_per_second(const struct pa_sample_spec *spec) {
     return spec->rate*pa_frame_size(spec);
 }
 
-pa_usec_t pa_bytes_to_usec(size_t length, const struct pa_sample_spec *spec) {
+pa_usec_t pa_bytes_to_usec(uint64_t length, const struct pa_sample_spec *spec) {
     assert(spec);
 
     return (pa_usec_t) (((double) length/pa_frame_size(spec)*1000000)/spec->rate);
